[feat] 카테고리 서비스 생성 #55
- CategoryService.java : 카테고리의 CRUD를 담당하는 서비스 메서드 생성했습니다.
